\name{CTTShiny-package}
\alias{CTTShiny-package}
\alias{CTTShiny}
\docType{package}
\title{
\packageTitle{CTTShiny}
}
\description{
\packageDescription{CTTShiny}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CTTShiny}
\packageIndices{CTTShiny}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CTTShiny}

Maintainer: \packageMaintainer{CTTShiny}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
