% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CTTShiny.R
\name{startCTT}
\alias{startCTT}
\title{This function will start CTTShiny}
\usage{
startCTT()
}
\value{
Nothing
}
\description{
An interactive Shiny application for running a CTT analysis.
}
\details{
Start CTTShiny

This starts the Shiny application CTT
}
\examples{
\dontrun{
library(shiny)
startCTT()
}
}
\keyword{CTT}

